Remove exhibitionAccessContent toggle
