+++ common/views/components/Footer/Footer.Nav.tsx
@@ -9,7 +9,6 @@ import Space from '@weco/common/views/components/styled/Space';
 
 const NavList = styled.ul<{
   $isInline?: boolean;
-  $exhibitionAccessContent?: boolean;
 }>`
   list-style-type: none;
   display: inline-block;
@@ -23,29 +22,8 @@ const NavList = styled.ul<{
   ${props =>
     props.$isInline &&
     `
-
-    ${
-      props.$exhibitionAccessContent
-        ? `
       li {
         display: inline-block;
-      }
-    `
-        : `
-    display: flex;
-    flex-direction: column;
-
-    ${props.theme.media('large')(`
-      flex-direction: row;
-
-      li:first-child a {
-        padding-top: 8px;
-      }
-    `)}
-    `
-    }
-
-      li {
         margin-right: 1.1rem;
 
         &:last-child {
@@ -108,23 +86,17 @@ const FooterNav = ({
   type,
   ariaLabel,
   isInline,
-  exhibitionAccessContent,
 }: {
   type: 'InternalNavigation' | 'PoliciesNavigation';
   ariaLabel: string;
   isInline?: boolean;
-  exhibitionAccessContent?: boolean;
 }): ReactElement => {
   const itemsList =
     type === 'PoliciesNavigation' ? PoliciesNavigation : InternalNavigation;
 
   return (
     <nav style={{ display: 'flex' }} aria-label={ariaLabel}>
-      <NavList
-        aria-label="Footer navigation"
-        $isInline={isInline}
-        $exhibitionAccessContent={exhibitionAccessContent}
-      >
+      <NavList aria-label="Footer navigation" $isInline={isInline}>
         {itemsList.map((link, i) => {
           // ID for Javascript-less users who tried to click on the Burger menu and will get redirected here
           const isBurgerMenuLink = type === 'InternalNavigation' && i === 0;
