+++ content/webapp/components/Exhibition/Exhibition.tsx
@@ -22,7 +22,6 @@ import {
   ExhibitionHighlightToursDocument,
   ExhibitionTextsDocument,
 } from '@weco/common/prismicio-types';
-import { useToggles } from '@weco/common/server-data/Context';
 import linkResolver from '@weco/common/services/prismic/link-resolver';
 import { font } from '@weco/common/utils/classnames';
 import { isFuture, isPast } from '@weco/common/utils/dates';
@@ -160,9 +159,7 @@ function getPlaceObject(
     }
   );
 }
-function getAccessibilityItems(
-  exhibitionAccessContent: boolean | undefined
-): ExhibitionItem[] {
+function getAccessibilityItems(): ExhibitionItem[] {
   const accessibilityItems: {
     description: prismic.RichTextField;
     icon: IconSvg;
@@ -208,38 +205,26 @@ function getAccessibilityItems(
       icon: accessibility,
     },
   ];
-  if (exhibitionAccessContent) {
-    return accessibilityItems.filter(item => {
-      if (item.description[0] && 'text' in item.description[0]) {
-        return item.description[0]?.text !== a11y.largePrintGuides;
-      } else {
-        return true;
-      }
-    });
-  } else {
-    return accessibilityItems.filter(item => {
-      if (item.description[0] && 'text' in item.description[0]) {
-        return (
-          item.description[0]?.text !== a11y.bsl &&
-          item.description[0]?.text !== a11y.accessResources
-        );
-      } else {
-        return true;
-      }
-    });
-  }
+
+  return accessibilityItems.filter(item => {
+    if (item.description[0] && 'text' in item.description[0]) {
+      return (
+        item.description[0]?.text !== a11y.bsl &&
+        item.description[0]?.text !== a11y.accessResources
+      );
+    } else {
+      return true;
+    }
+  });
 }
 
-export function getInfoItems(
-  exhibition: ExhibitionType,
-  exhibitionAccessContent?: boolean
-): ExhibitionItem[] {
+export function getInfoItems(exhibition: ExhibitionType): ExhibitionItem[] {
   return [
     getUpcomingExhibitionObject(exhibition),
     getadmissionObject(),
     getTodaysHoursObject(),
     getPlaceObject(exhibition),
-    ...getAccessibilityItems(exhibitionAccessContent),
+    ...getAccessibilityItems(),
   ].filter(isNotUndefined);
 }
 
@@ -266,7 +251,6 @@ const Exhibition: FunctionComponent<Props> = ({
 }) => {
   type ExhibitionOf = (ExhibitionType | EventBasic)[];
 
-  const { exhibitionAccessContent } = useToggles();
   const [exhibitionOfs, setExhibitionOfs] = useState<ExhibitionOf>([]);
   const [exhibitionAbouts, setExhibitionAbouts] = useState<ExhibitionAbout[]>(
     []
@@ -554,13 +538,10 @@ const Exhibition: FunctionComponent<Props> = ({
       // We hide contributors as we show them further up the page
       hideContributors={true}
     >
-      {exhibitionAccessContent && exhibition.uid !== 'being-human' ? (
+      {exhibition.uid !== 'being-human' ? (
         <>
           {exhibition.end && !isPast(exhibition.end) && (
-            <InfoBox
-              title="Visit us"
-              items={getInfoItems(exhibition, exhibitionAccessContent)}
-            />
+            <InfoBox title="Visit us" items={getInfoItems(exhibition)} />
           )}
 
           {(exhibitionOfs.length > 0 || pages.length > 0) && (
@@ -736,10 +717,7 @@ const Exhibition: FunctionComponent<Props> = ({
 
           {exhibition.end && !isPast(exhibition.end) && (
             <Space $v={{ size: 'l', properties: ['margin-bottom'] }}>
-              <InfoBox
-                title="Visit us"
-                items={getInfoItems(exhibition, exhibitionAccessContent)}
-              >
+              <InfoBox title="Visit us" items={getInfoItems(exhibition)}>
                 <AccessibilityServices>
                   For more information, please visit our{' '}
                   <a href={`/visit-us/${prismicPageIds.access}`}>
